sysutils/u-boot-*: Add patch to fix MMC ordering
For FreeBSD we use a dtb overlay that switches between the sdhost and
the sdhci controller for sd card.
Recent changes in u-boot device model changed device numbering and the
sd card controller became "mmc2" for our setup which isn't accounted for
in the possible boot devices.
Patch isn't accepted yet upstream but bring this patch in for our ports anyway.