x11-drivers/xf86-video-vesa: Add patch for ignoring if kernel as a driver
The vesa driver checks if the kernel as a driver attached to the pci device.
This used to work before df10dcefa427 ("devel/libpciaccess: Update to 0.17")
because we had a patch in libpciaccess that always said that the kernel didn't
had any driver attached. This is obviously not a correct way.
The problem is that vgapci is always attached for us so for pci video devices
we always have a driver attached.
Ignoring the check in xf86-video-vesa seems the best way for us.

PR:	270509
Sponsored by:	Beckhoff Automation GmbH & Co. KG