x11-servers/xorg-server: add missing features for PRIME
Modern PRIME X11 setups autoconfigure themselves, which internally
uses the udev and udev_kms options to do so. These options are required
for X to find the available DRM devices. This is needed for PRIME on
FreeBSD.

This is also adds the DRM platform code, taken from X commit:

https://gitlab.freedesktop.org/xorg/xserver/-/commit/5468123822bbe8cd0a0abe07bbbd3a9cb14ed2cb

Differential Revision:	https://reviews.freebsd.org/D39886

Which is in master but not yet in any releases. This adds the platform
code to the build for DRM so that PRIME setups can work.