x11-servers/xorg-server: Add upstream patch for xkbcomp directory
Add a patch commited upstream but not present in any releases (except
Xwayland ones) as xorg-server releases stopped a while ago. This adds
a fallback to use $XDG_RUNTIME_DIR for xkbcomp result file if the user
can't write to the configured directory, which is unlikely for xorg-server
as it's SUID but it is likely for xephyr or xvfb.
While here use the meson default for xkb_dir as it's the same as the one we
provide and use also the default for xkb_output_dir which defaults to
xkb_dir/compiled which is a link to /var/lib/xkb (setup by the xkeyboard-config
port).

Sponsored by:	Beckhoff Automation GmbH & Co. KG