multimedia/x265: update to 3.6
The latest edition provides fixes for the 32-bit ARM, and drops the need
for gcc/binutils on the 64-bit ARM (aarch64). Many thanks to fuz@ for
providing remote access to the two ARM-systems.

The libvmaf-support remains broken, because our port of the dependency
has been upgraded to 3.0.0, which is only supported by x265-4.0.  Not
removing this option for the sake of those, who might have kept the
older vmaf on their systems.

The ASM is no longer an option -- it was always a requirement on some
platforms, and is very useful on others.

PR:		281680